# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BackBet API",
    "version": "1.0.0",
    "description": "API de Apostas - Backend",
    "contact": {
      "name": "BackBet Team",
      "url": "https://github.com/dioneyfroes-coder/backBet"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Desenvolvimento"
    },
    {
      "url": "https://api.backbet.com/api",
      "description": "Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token ou User ID em desenvolvimento"
      }
    },
    "schemas": {
...
      "PlaceBetRequest": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "marketId": {
            "type": "string",
            "example": "market-123"
          },
          "oddId": {
            "type": "string",
            "example": "odd-456"
          },
          "amount": {
            "type": "number",
            "example": 50,
            "minimum": 0.01
          },
          "type": {
            "type": "string",
            "enum": [
              "SINGLE",
              "MULTIPLE"
            ],
            "example": "SINGLE"
          },
          "currency": {
            "type": "string",
            "enum": [
              "BRL",
              "USD",
              "EUR"
            ],
            "example": "BRL"
          }
        },
        "required": [
          "eventId",
          "marketId",
          "oddId",
          "amount"
        ]
      },
      "CancelBetRequest": {
        "type": "object",
        "properties": {
          "betId": {
            "type": "string",
            "format": "uuid",
            "example": "7d9f6c2b-1b2a-4a8b-9c0d-123456789abc"
          },
          "reason": {
            "type": "string",
            "example": "Erro no mercado"
          }
        },
        "required": [
          "betId"
        ]
      },
      "BetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "7d9f6c2b-1b2a-4a8b-9c0d-123456789abc"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "cadaeb28-c7f7-425b-91f7-73a27141ae49"
          },
          "eventId": {
            "type": "string",
            "format": "uuid",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "marketId": {
            "type": "string",
            "example": "market-123"
          },
          "amount": {
            "type": "number",
            "example": 50
          },
          "odds": {
            "type": "number",
            "example": 1.85
          },
          "potentialReturn": {
            "type": "number",
            "example": 92.5
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "WON",
              "LOST",
              "CANCELED"
            ],
            "example": "PENDING"
          },
          "type": {
            "type": "string",
            "example": "SINGLE"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-14T23:30:00.000Z"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BetListResponse": {
        "type": "object",
        "properties": {
          "bets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BetResponse"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/bets/event/{eventId}": {
      "get": {
        "tags": [
          "Bets"
        ],
        "summary": "Lista apostas de um evento",
        "parameters": [
          {
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de apostas do evento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bets": {
      "post": {
        "tags": [
          "Bets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Coloca uma nova aposta",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceBetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Aposta criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          }
        }
      }
    },
    "/api/bets/{betId}/cancel": {
      "post": {
        "tags": [
          "Bets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Cancela uma aposta pendente",
        "parameters": [
          {
            "in": "path",
            "name": "betId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Aposta cancelada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Não pode cancelar"
          }
        }
      }
    },
    "/api/bets/me": {
      "get": {
        "tags": [
          "Bets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Lista apostas do usuário autenticado",
        "responses": {
          "200": {
            "description": "Lista de apostas do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/games/coin-flip/play": {
      "post": {
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Executa uma rodada de cara ou coroa",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "choice": {
                    "type": "string",
                    "enum": [
                      "HEADS",
                      "TAILS"
                    ]
                  },
                  "wager": {
                    "type": "number",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado da rodada"
          }
        }
      }
    },
    "/api/games/coin-flip/history": {
      "get": {
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Histórico de cara ou coroa do usuário",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico recente"
          }
        }
      }
    },
...
  }
}